cleaned up JS code and fixed defaulting of mixer after playout
