+++ frontend/src/constants.js
@@ -1,0 +1,29 @@
+export const CSS_CLASSES = {
+  EDIT_ONLY: "edit-only",
+  LIVE_ONLY: "live-only",
+  IS_LIVE: "is-live",
+  MODE_LIVE: "mode-live",
+  FIELD_ROW: "field-row",
+  WIDGET_CARD: "widget-card",
+  CUSTOM_FIELDS: "custom-fields-container",
+  DELETE_BTN: "delete-btn",
+  ACTION_BTN: "action-btn",
+  ADD_FIELD_BTN: "add-field-btn",
+};
+
+export const SELECTORS = {
+  TOGGLE_MODE_BTN: "#toggle-mode-btn",
+  ADD_WIDGET_BTN: "#add-widget-btn",
+  FIELD_KEY: ".f-key",
+  FIELD_TYPE: ".f-type",
+  FIELD_ID: ".f-id",
+  FIELD_SOURCE: ".f-source",
+  LIVE_KEY_DISPLAY: ".live-key-display",
+  LIVE_VALUE_DISPLAY: ".live-value-display",
+};
+
+export const FIELD_TYPES = {
+  STRING: "string",
+  INT: "int",
+  FLOAT: "float",
+};
