+++ frontend/src/state.js
@@ -1,0 +1,8 @@
+/**
+ * Shared application state.
+ * grid is set during initializeApp() once GridStack is ready.
+ */
+export const AppState = {
+  isLiveMode: false,
+  grid: null,
+};
